backend/requirements.txt
flask
flask-cors
google-generativeai>=0.7.2
python-dotenv
gunicorn
requests